Add GUI alerts, add new NDX PVs
